requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.1
altgraph==0.17.4
attrs==25.1.0
backports.tarfile==1.2.0
bcrypt==4.2.1
blinker==1.9.0
Brotli==1.1.0
cachelib==0.13.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
customtkinter==5.2.2
darkdetect==0.8.0
Flask==3.1.0
Flask-Compress==1.17
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
geographiclib==2.0
greenlet==3.1.1
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.5
keyring==25.6.0
Mako==1.3.8
MarkupSafe==3.0.2
maxminddb==2.6.3
more-itertools==10.6.0
msgspec==0.19.0
multidict==6.1.0
packaging==24.2
paramiko==3.5.1
pefile==2023.2.7
pillow==11.1.0
propcache==0.2.1
pycparser==2.22
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
PyJWT==2.10.1
PyMySQL==1.1.1
PyNaCl==1.5.0
python-dotenv==1.0.1
pywin32==308
pywin32-ctypes==0.2.3
requests==2.32.3
SQLAlchemy==2.0.37
typing_extensions==4.12.2
urllib3==2.3.0
waitress==3.0.2
Werkzeug==3.1.3
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
